Reject stateful runtime defaults
